requirements.txt
streamlit==1.23.0
streamlit-folium==0.11.1
pandas==1.5.3
inflection==0.5.1
plotly==5.14.1
folium==0.14.0
haversine==2.8.0
Pillow==9.5.0